Create a small helper for the once initialization
The lock and initialized parameters cannot be located in the common code
because they must be unique per implementation. Technically the lock
could be shared, but the initialized toggle must be unique since it
needs to allow each implementation to be initialized. Sharing the lock
would create unnecessary contention since the implementation specific
fields do not overlap.